Fix query/predicate circular deps
Fix query/predicate circular deps

Closes #795
