+++ packages/core/src/query.ts
@@ -1,5 +1,6 @@
 import { DocumentQuery } from '.'
 import { Doc } from './classes'
+import { getObjectValue } from './objvalue'
 import { createPredicates, isPredicate } from './predicate'
 import { SortingQuery } from './storage'
 
@@ -16,7 +17,7 @@ export function findProperty (objects: Doc[], propertyKey: string, value: any): 
   }
   const result: Doc[] = []
   for (const object of objects) {
-    const val = getNestedValue(propertyKey, object)
+    const val = getObjectValue(propertyKey, object)
     if ((val === value) || isArrayValueCheck(val, value)) {
       result.push(object)
     }
@@ -26,48 +27,6 @@ export function findProperty (objects: Doc[], propertyKey: string, value: any): 
 
 function isArrayValueCheck<T, P> (val: T, value: P): boolean {
   return Array.isArray(val) && !Array.isArray(value) && val.includes(value)
-}
-/**
- * @public
- */
-export function getNestedValue (key: string, doc: Doc): any {
-  // Check dot notation
-  if (key.length === 0) {
-    return doc
-  }
-  key = key.split('\\$').join('$')
-  const dots = key.split('.')
-  // Replace escapting, since memdb is not escape keys
-
-  // We have dots, so iterate in depth
-  let pos = 0
-  let value = doc as any
-  for (const d of dots) {
-    if (Array.isArray(value) && isNestedArrayQuery(value, d)) {
-      // Array and d is not an indexed field.
-      // So return array of nested values.
-      return getNestedArrayValue(value, dots.slice(pos).join('.'))
-    }
-    value = value?.[d]
-    pos++
-  }
-  return value
-}
-
-function isNestedArrayQuery (value: any, d: string): boolean {
-  return Number.isNaN(Number.parseInt(d)) && value?.[d as any] === undefined
-}
-
-function getNestedArrayValue (value: any[], name: string): any[] {
-  const result = []
-  for (const v of value) {
-    result.push(...arrayOrValue(getNestedValue(name, v)))
-  }
-  return result
-}
-
-function arrayOrValue (vv: any): any[] {
-  return Array.isArray(vv) ? vv : [vv]
 }
 
 /**
@@ -97,7 +56,7 @@ function getSortingResult (aValue: any, bValue: any): number {
 }
 
 function getValue (key: string, obj: any): any {
-  let value = getNestedValue(key, obj)
+  let value = getObjectValue(key, obj)
   if (typeof value === 'object') {
     value = JSON.stringify(value)
   }
